changed return value to be conform with input value. added PID controller

+++ gams/src/main/java/eu/arrowhead/gams/controller/BangBangController.java
@@ -33,9 +33,9 @@ public class BangBangController implements TriggeredController
     }
 
     @Override
-    public Integer evaluate(final Long inputValue, final Supplier<Long> targetValue)
+    public Long evaluate(final Long inputValue, final Supplier<Long> targetValue)
     {
-        int value = (int) (targetValue.get() - inputValue);
+        long value = (targetValue.get() - inputValue);
 
         if (inverse)
         { value = value * (-1); }
